update the organization of the classes to get ready for dnd
